Reorg PromotionPickerScreen into own package
