+++ core/src/com/unciv/ui/screens/worldscreen/unit/UnitTable.kt
@@ -24,8 +24,8 @@ import com.unciv.ui.screens.basescreen.BaseScreen
 import com.unciv.ui.screens.civilopediascreen.CivilopediaCategories
 import com.unciv.ui.screens.civilopediascreen.CivilopediaScreen
 import com.unciv.ui.screens.pickerscreens.CityRenamePopup
-import com.unciv.ui.screens.pickerscreens.PromotionPickerScreen
-import com.unciv.ui.screens.pickerscreens.UnitRenamePopup
+import com.unciv.ui.screens.pickerscreens.promotion.PromotionPickerScreen
+import com.unciv.ui.screens.pickerscreens.promotion.UnitRenamePopup
 import com.unciv.ui.screens.worldscreen.WorldScreen
 
 class UnitTable(val worldScreen: WorldScreen) : Table() {
