+++ core/src/com/unciv/ui/screens/worldscreen/unit/actions/UnitActions.kt
@@ -28,7 +28,7 @@ import com.unciv.ui.components.Fonts
 import com.unciv.ui.popups.ConfirmPopup
 import com.unciv.ui.popups.hasOpenPopups
 import com.unciv.ui.screens.pickerscreens.ImprovementPickerScreen
-import com.unciv.ui.screens.pickerscreens.PromotionPickerScreen
+import com.unciv.ui.screens.pickerscreens.promotion.PromotionPickerScreen
 
 object UnitActions {
